Move refresh to a two-tier queuing.
With only one tier of queuing, it sometimes happens that a first try to
fetch the item information is done, returns a stub, then by the time the
second try happens the information has been lost again. With a queue of
15K+ items at this point, this means nothing gets processed and we loop
infinitely.

Apply a bit of heuristics and do the processing in two tiers. Retry the
most recent items a few times before putting them back at the end. This
speeds up the refreshing a bit and also makes it way less likely to
spiral into infinite looping.